added logic
