+++ app/src/main/java/org/myspecialway/ui/alarms/AlarmsReceiver.kt
@@ -1,0 +1,46 @@
+package org.myspecialway.ui.alarms
+
+import android.annotation.SuppressLint
+import android.arch.persistence.room.Room
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import com.github.ajalt.timberkt.d
+import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
+import org.myspecialway.common.getRemainingAlarmsForToday
+import org.myspecialway.data.local.Database
+import org.myspecialway.di.LocalProperties
+import org.myspecialway.ui.agenda.ScheduleModel
+import org.myspecialway.ui.agenda.mapScheduleRenderModel
+
+
+class RestartReceiver : BroadcastReceiver() {
+    @SuppressLint("CheckResult")
+    override fun onReceive(context: Context?, intent: Intent?) {
+
+        val list = getLocalSchedule(context)
+
+        // get the list of schedule
+        list.subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .map { it.data.classById.schedule }
+                .subscribe({ scheduleList ->
+
+                    val render = scheduleList
+                            .map { mapScheduleRenderModel(it) }
+
+                    val remainingAlarms = render.toMutableList()
+                            .getRemainingAlarmsForToday()
+
+                    AlarmJob.scheduleJobs(remainingAlarms)
+                }, { d { "can't get the schedule list" } })
+    }
+
+    private fun getLocalSchedule(context: Context?): Single<ScheduleModel> =
+            Room.databaseBuilder(context!!, Database::class.java, LocalProperties.DATABASE_NAME)
+                    .build()
+                    .localDataSourceDAO()
+                    .loadSchedule()
+}
